Revert "ITEP-69567: Upgrade openvino version to 2025.1 (#80)"
Method to download open model zoo models has changed from previous
version. Will break build for a fresh repo.
