+++ tests/sscape_tests/detector/conftest.py
@@ -18,10 +18,12 @@ import uuid
 import numpy as np
 from PIL import Image
 from PIL import ImageDraw
+from unittest.mock import patch
 from unittest.mock import MagicMock
 
 from percebro import videosource, detector, detector_tesseract, detector_atag, detector_motion, \
-  detector_geti, detector_ocr, detector_pose
+  detector_geti, detector_ocr
+from scene_common import camera
 from scene_common.timestamp import get_epoch_time
 from tests.sscape_tests.detector.config import ovms_retail_model, ovms_hpe_model
 import tests.common_test_utils as common
@@ -163,7 +165,7 @@ def start_inference(detector_object, input_data, preprocessed_data):
 def pose_estimator():
   """! Creates a PoseEstimator object for this module """
 
-  poseEstimator = detector_pose.PoseEstimator()
+  poseEstimator = detector.PoseEstimator()
   poseEstimator.setParameters(pose_model, device, plugin, threshold, openvino_cores)
 
   return poseEstimator
@@ -476,7 +478,7 @@ def ovms_detector():
 def ovms_hpe():
   """! Creates a OVMS HPE object for this module """
 
-  detector_obj = detector_pose.PoseEstimator(distributed=detector.Distributed.OVMS)
+  detector_obj = detector.PoseEstimator(distributed=detector.Distributed.OVMS)
   detector_obj.setParameters(ovms_hpe_model, device, plugin, threshold, openvino_cores)
 
   return detector_obj
